Update after-build-package task test
